feat: infinite query 90%
