+++ components/common/ProfileImage.tsx
@@ -45,4 +45,5 @@ const ProfileImageStyle = styled(Image)`
   width: 100%;
   height: 100%;
   object-fit: cover;
+  cursor: pointer;
 `;
